refactor route handler
