+++ core/RouteHandler.ts
@@ -1,7 +1,26 @@
+import {Router} from 'express';
+import {ControllerRegistration} from '../controllers/ControllerDecorator';
 /**
  * TODO
  */
 export interface RouteHandler {
+    /**
+     * TODO
+     *
+     */
+    addRoutes(controllerRegistration: ControllerRegistration, url: string): void;
+
+    /**
+     * TODO
+     *
+     * @param router
+     */
+    registerRoutes(router: Router): Router;
+
+    /**
+     * TODO
+     */
+    resetRoutes(): void;
 }
 
 /**
